NetworkManager updated with lobby
Added lobby functionality to NetworkManager.cs
Now server has a list of player names as they connect.
Players have to insert ipaddress and port but once we get an actual
server that can host games we won't have to do that anymore (unless we
want to use Unity's master list).