DFS Maze now tries to place exit away from start
A distance method was added to Square.cs to compare the distance between
Squares A and B. The DepthFirstMaze generator uses this method to
determine an ample distance away from the start square to place the maze
exit. Testing has shown this to work for mazes of ample sizes. A
quadrant based method tries to catch any cases where the distance method
fails to find an exit ample distance away for smaller maze sizes.