Added simple powerup aura on item pickup
When the player picks up a powerup orb, a particle system is created and
added to the PlayerAvatar item hierarchy. The aura color will generally
match the color of the item that spawned it - in this case it will match
the color of any powerup. The particle system destroy itself shortly
after the player picks it up.