Attempted controller bug fix on start screen
All objects on the start screen were lowered on the y-axis by 1 unit.
This matches the Player's environment in the game itself. As a result of
lowering each object, the controller bug where a player is required to
place their finger on the middle of the screen to move the joystick
seems to disappear. However, the floor texture on the start screen and
the multiplayer lobby fails to load until the game is completed once and
the main menu is loaded again.
This build has only been tested on my Android phone. Further testing on
multiple devices is necessary to substantiate claim