More Lerp and Fog of War changes
Fog of War is set to spawn one unit lower in the Y-direction if the
player is a client.
Without this change, the client screen would not illuminate as well as
the host's screen.
NormalizePosition in NetworkCharacter changed to an abstract method.
Without the change, NetworkMonsters would spawn at a height too high on
client screens
Also, a check was added to NetworkUser to only normalize if the player
is the client
Without this check, all clients would move around at a height too high
on the host's screen