Loading Screen and End Conditions updated
Loading screen now uses updated Unity UI features and looks a lot
smoother
End conditions updated to display the button that transitions to the end
of the game if any player reaches the end. However, for every player
that didn't reach the end, they're prompted with a "You Lose!" dialog.