Network Attack Issue Patch
Untested
Although, I don't see why this won't work.
Previous attempts to fix the network attack issue where if the player
attacks, then all players attack did not modify the Unity Editor section
of the code. This commit tries to block that section if the attacking
player's network view does not belong to the player